src/pyproject.toml
[tool.poetry]
name = "base-fastapi-project"
version = "0.1.0"
description = ""
authors = ["Anton Kholkin <kholkin.anthony@gmail.com>"]
readme = "../README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.0"
alembic = "^1.11.1"
inflection = "^0.5.1"
uvicorn = "^0.22.0"
pydantic = {extras = ["dotenv"], version = "^2.1.1"}
pydantic-settings = "^2.0.2"
sqlalchemy = {extras = ["mypy"], version = "^2.0.15"}
asyncpg = "^0.27.0"
greenlet = "^2.0.2"
httpx = "^0.24.1"
python-multipart = "^0.0.6"
loguru = "^0.7.0"
authlib = "^1.2.1"
itsdangerous = "^2.1.2"
pytz = "^2024.2"

[tool.poetry.group.dev.dependencies]
ruff = "^0.1.7"
bump-pydantic = "^0.6.1"
pytest = "^8.1.1"
pytest-asyncio = "^0.21.0"
pytest-factoryboy = "^2.5.1"
pytest-mock = "^3.14.0"
pytest-repeat = "^0.9.3"
types-redis = "^4.6.0.0"
types-pytz = "^2024.1.0.20240417"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120

